refactor: dto 및 model dto 디렉토리로 이동
